Fixed warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@90208 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/objostrasn.cpp
@@ -329,7 +329,7 @@ void CObjectOStreamAsn::WriteBitString(const CBitString& obj)
         CBitString::statistics st;
         obj.calc_stat(&st);
         char* buf = (char*)malloc(st.max_serialize_mem);
-        unsigned int len = bm::serialize(obj, (unsigned char*)buf, tmp_block);
+        size_t len = bm::serialize(obj, (unsigned char*)buf, tmp_block);
         WriteBytes(buf,len);
         free(buf);
         bm::aligned_free(tmp_block);
